Implemented getUserProfile method in SocialServiceImpl, modified User Entity, ProfileDTO and AppConfig

+++ src/main/java/com/laa66/statlyapp/DTO/ProfileDTO.java
@@ -21,5 +21,5 @@ public class ProfileDTO {
     List<Track> topTracks;
     List<Artist> topArtists;
     List<PlaylistInfo> userPlaylists;
-    int mainstream;
+    double mainstream;
 }
